correction typage
(Logical change 1.1177)


git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@11322 ee107e0b-c2fa-0310-842a-ec2723fb34f4

+++ pm2/pm2/marcel/include/asm-alpha/linux_bitops.h
@@ -392,10 +392,10 @@ static inline int ma_ffs(int word)
  * fls: find last bit set.
  */
 #section marcel_functions
-static inline int ma_fls(int word);
+static inline unsigned long ma_fls(unsigned long word);
 #section marcel_inline
 #if defined(__alpha_cix__) && defined(__alpha_fix__)
-static inline int ma_fls(int word)
+static inline unsigned long ma_fls(unsigned long word)
 {
 	return 64 - __kernel_ctlz(word & 0xffffffff);
 }
